Rename PluginItem to VisualizationItem
